refactor(automations): extract dispatch_run; move manual adapter under triggers/manual/dispatch.py
