+++ surfsense_backend/app/automations/triggers/manual/__init__.py
@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
+from .dispatch import dispatch_manual_run
 from .params import ManualTriggerParams
 
-__all__ = ["ManualTriggerParams"]
+__all__ = ["ManualTriggerParams", "dispatch_manual_run"]
 
 # Side-effect: register on the triggers store.
 from . import definition  # noqa: E402, F401
