+++ surfsense_backend/app/automations/triggers/manual/dispatch.py
@@ -1,25 +1,18 @@
-"""Manual ``Run now`` dispatch: validate inputs, snapshot the definition, enqueue."""
+"""Manual ``Run now`` dispatch adapter: load + guard, then call generic dispatch."""
 
 from __future__ import annotations
 
 from typing import Any
 
-import jsonschema
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.automations.dispatch import DispatchError, dispatch_run
 from app.automations.persistence.enums.automation_status import AutomationStatus
-from app.automations.persistence.enums.run_status import RunStatus
 from app.automations.persistence.enums.trigger_type import TriggerType
 from app.automations.persistence.models.automation import Automation
 from app.automations.persistence.models.run import AutomationRun
 from app.automations.persistence.models.trigger import AutomationTrigger
-from app.automations.schemas.definition.envelope import AutomationDefinition
-from app.automations.tasks.execute_run import automation_run_execute
-
-
-class DispatchError(Exception):
-    """A manual dispatch could not proceed (missing trigger, invalid inputs, ...)."""
 
 
 async def dispatch_manual_run(
@@ -28,7 +21,7 @@ async def dispatch_manual_run(
     automation_id: int,
     payload: dict[str, Any] | None,
 ) -> AutomationRun:
-    """Validate, snapshot, persist, and enqueue an ``AutomationRun``."""
+    """Find the automation + its enabled manual trigger, then run the generic dispatch."""
     automation = await _load_automation(session, automation_id)
     if automation is None:
         raise DispatchError(f"automation {automation_id} not found")
@@ -38,39 +31,18 @@ async def dispatch_manual_run(
             f"automation {automation_id} is {automation.status.value}, not active"
         )
 
-    try:
-        definition = AutomationDefinition.model_validate(automation.definition)
-    except Exception as exc:
-        raise DispatchError(f"invalid automation definition: {exc}") from exc
-
     trigger = await _find_manual_trigger(session, automation_id)
     if trigger is None:
         raise DispatchError(
             f"automation {automation_id} has no enabled manual trigger"
         )
 
-    resolved_inputs = _validate_inputs(definition, payload or {})
-    snapshot = definition.model_dump(mode="json", by_alias=True)
-
-    run = AutomationRun(
-        automation_id=automation_id,
-        trigger_id=trigger.id,
-        status=RunStatus.PENDING,
-        definition_snapshot=snapshot,
-        trigger_payload=payload,
-        resolved_inputs=resolved_inputs,
-        step_results=[],
-        artifacts=[],
+    return await dispatch_run(
+        session=session,
+        automation=automation,
+        trigger=trigger,
+        payload=payload,
     )
-    session.add(run)
-    await session.commit()
-    await session.refresh(run)
-
-    automation_run_execute.apply_async(
-        args=[run.id],
-        time_limit=definition.execution.timeout_seconds,
-    )
-    return run
 
 
 async def _load_automation(
@@ -93,15 +65,3 @@ async def _find_manual_trigger(
         .limit(1)
     )
     return (await session.execute(stmt)).scalar_one_or_none()
-
-
-def _validate_inputs(
-    definition: AutomationDefinition, payload: dict[str, Any]
-) -> dict[str, Any]:
-    if definition.inputs is None or not definition.inputs.schema_:
-        return {}
-    try:
-        jsonschema.validate(instance=payload, schema=definition.inputs.schema_)
-    except jsonschema.ValidationError as exc:
-        raise DispatchError(f"inputs: {exc.message}") from exc
-    return payload
